feat: iniciando funcionalidade de ocultar categorias padrão

+++ Repositories/CategoriaRepository.cs
@@ -36,7 +36,8 @@ namespace fin_api.Repositories
         public async Task DeleteAsync(string id)
         {
             var categoria = await GetByIdAsync(id);
-            if (categoria != null)
+            if (categoria != null && categoria.IsDefault) await HiddenCategory(categoria);
+            if (categoria != null && !categoria.IsDefault)
             {
                 _context.Categories.Remove(categoria);
                 await _context.SaveChangesAsync();
@@ -49,6 +50,32 @@ namespace fin_api.Repositories
                 .AnyAsync(c => (c.UserId == userId || c.IsDefault ) && c.Name.ToLower() == name.ToLower());
         }
 
+        public async Task<bool> IsCategoryHiddenAsync(string userId, string name)
+        {
+            var category = await _context.Categories
+                .FirstOrDefaultAsync(c => c.Name.ToLower() == name.ToLower() && c.IsDefault);
 
+            if (category == null || !category.IsDefault) return false;
+
+            return await _context.UserHiddenCategories
+                .AnyAsync(uhc => uhc.UserId == userId && uhc.CategoryId == category.Id);
+        }
+
+        private async Task HiddenCategory(Categoria categoria)
+        {
+            var hiddenCategory = new UserHiddenCategory { UserId = categoria.UserId, CategoryId = categoria.Id };
+            await _context.UserHiddenCategories.AddAsync(hiddenCategory);
+            await _context.SaveChangesAsync();
+        }
+        public async Task ShowHiddenCategory(Categoria categoria)
+        {
+            var hiddenCategory = await _context.UserHiddenCategories
+                .FirstOrDefaultAsync(uhc => uhc.UserId == categoria.UserId && uhc.CategoryId == categoria.Id);
+            if (hiddenCategory != null)
+            {
+                _context.UserHiddenCategories.Remove(hiddenCategory);
+                await _context.SaveChangesAsync();
+            }
+        }
     }
 }
